requirements.txt
openai>=1.65.4
transformers>=4.46.3
tiktoken>=0.9.0
numpy>=1.21.0
pyyaml>=5.4.0
requests>=2.26.0 